+++ app/src/minimal/java/io/homeassistant/companion/android/sensors/ActivitySensorManager.kt
@@ -4,6 +4,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import io.homeassistant.companion.android.R
+import io.homeassistant.companion.android.common.sensors.SensorManager
 
 class ActivitySensorManager : BroadcastReceiver(), SensorManager {
